Get rid of installation token (for OSS installations)
This is done by being required to supply the grafana API URL as an
environment variable on the backend. Additionally, optionally an OnCall
API URL environment variable can be passed in to the frontend (this basically
allows completely skipping the need to configure anything).

- deduplicated a lot of the sync logic on the frontend + made
error message more useful and consistent
- Converted jest.config.js to jest.config.ts
- Upgraded jest from 27.5.1 to latest (29.2.3)
- Added tests for:
  - PluginConfigPage component
  - state/plugin and state/rootBaseStore functions
  - apps.grafana_plugin django app